Refactor Clean and Init tasks, move linking jobs to run in the background (#303)
### Description
This PR refactors some of the tasks. The goal is to reduce the number of jobs to be submitted, hence, computing and waiting time for the each cycle. `Clean` tasks are removed and the content is moved to their associated application tasks (e.g., `CleanHofX` content is moved to `HofX.csh`). `InitVariationals.csh` and `InitEnKF.csh` scripts for the YAML and model stage preparation stage are also removed and their content is moved to `PrepJEDI.csh` (only one preparation task is submitted). For RTPP, `Clean` and `Init` tasks and both removed and merged with `RTPP.csh`. Here, the `LinkExternalAnalysis` and `LinkWarmStartBackgrounds` tasks are set to run in the background instead of Casper. These tasks use very little memory and wall-time, then running them in the background should be fine (but let's monitor that).

### Issue closed
None

### Tests completed
#### Tier 1:
 - [ ] 3dvar_OIE120km_WarmStart
 - [ ] 3denvar_OIE120km_IAU_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [ ] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart
 - [x] getkf_OIE120km_WarmStart
 - [x] 4dhybrid_OIE120km_WarmStart
 - [ ] ForecastFromGFSAnalysesMPT

#### Tier 2:
- [x] 3dhybrid-allsky_O30kmIE60km_SpecifiedEnsemble_VarBC