Add variable mesh 60-3km for forecast (#304)
### Description
This PR adds specification for the variable mesh 60-3km. The variable `meshRatio` is introduced in order to get correct static, graph and info files for each application in a more generic way. The namelist is also made more configurable, adding high flexibility for the physical parameterizations selection. For this mesh, Thompson microphysics is used for the forecasts, then the `IRVISlandCoeff` is made configurable in order to use the `IGBP` coefficients for the HofX application (default is `USGS`) and the table for Thompson microphysics are added. `ForecastFromGFSAnalysesVariableMesh` scenario is added for this functionality.

At this moment, this mesh does not work fully for data assimilation (dual-resolution). Although many of the changes introduced here work for the DA, some additional modifications are still needed, which will be the subject of a future PR.

### Issue closed
None

### Tests completed
#### Tier 1:
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_IAU_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart
 - [x] getkf_OIE120km_WarmStart
 - [x] 4dhybrid_OIE120km_WarmStart
 - [x] ForecastFromGFSAnalysesMPT

#### Tier 3 (optional):
 - [x] ForecastFromGFSAnalysesVariableMesh