Add functionality for cloud direct insertion (SACA) (#305)
### Description
This PR adds the functionality to do cloud direct insertion (SACA) on cold start forecasts and then run forecasts from the SACA analysis (using the `ExtendedForecast` tasks). A new suite is added for this capability. The default is doing SACA analysis on the 0-hr forecast from a cold start forecast initialized from GFS analysis, but it can be done on the 6-hr as well by changing the variable `prevBgHR`. This variable is used to trigger the cold forecast at the indicated hour (i.e., at the same hour as the SACA analysis (e.g., 00Z), or 6 hours before (e.g., 18Z)). SACA analysis uses a similar approach as the `ExtendedForecast`, thus it does the SACA analysis at the specified mean times. Sea surface fields are updated from the GFS analysis during the `ExtendedForecast` step. The path for the build code is updated.

This PR also includes a fix for the `Concatenate` task so the verification of observations waits until the observations output have been concatenated.

### Issue closed
None

### Tests completed
#### Tier 1:
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_IAU_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] 4denvar_OIE120km_WarmStart
 - [x] 4dhybrid_OIE120km_WarmStart
 - [x] eda_OIE120km_WarmStart
 - [x] getkf_OIE120km_WarmStart
 - [x] ForecastFromGFSAnalysesMPT 

#### Tier 3 (optional):
 - [x] ForecastFromGFSAnalysesVariableMesh.yaml
 - [x] CloudDirectInsertion.yaml