Add scenario with all possible configuration for hybrid-3denvar 30km-60km warm start experiment (#316)
### Description
This PR adds a top scenario yaml file for a single experiment (3dhybrid_O30kmIE60km_SpecifiedEnsemble_VarBC) containing most possible user configurable variables. This scenario can help us to better locate the default paths and variable values we are using in each section for this experiment. The section for the extendedForecast is included for reference but turned off.

### Issue closed
None

### Tests completed
#### Tier 3 (optional):
 - [x] 3dhybrid_O30kmIE60km_SpecifiedEnsemble_VarBC_allConfig