Some clean-up and folder specification for external analyses and first bkg (#322)
### Description
This small PR is to adjust the folder specification for the external analyses and first background for the 60-3km variable mesh to use the ones created with new topography files and land use option. It also moves back some scenario yaml configurations and fixes a typo in the CloudDirectInsertion suite.

### Issue closed
None

### Tests completed
Files where updated and tested.