Add new suite for cycling SACA with the variational data assimilation (#328)
### Description
This PR adds a new suite to apply SACA before or after the cycling DA. A variable is added to control this new capability (`runCyclingSaca`) and the variable `prevBgHR` is used to control whether apply SACA before or after in order to link the right background file for SACA and DA and the initial condition for the forecast. The default is cycling SACA after DA (`runCyclingSaca`: True). A new scenario is added for this capability. 

The path for the graphics is updated to include recent PR to plot Gradient Norm values.

### Issue closed
None

### Tests completed
#### Tier 1:
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_IAU_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart
 - [x] getkf_OIE120km_WarmStart
 - [x] ForecastFromGFSAnalysesMPT

#### Tier 2:
 - [x] 3dhybrid_O30kmIE60km_SpecifiedEnsemble_VarBC_CloudInsertion
 - [x] CloudDirectInsertion