Remove interpolation type (#336)
### Description
This PR removes the `interpolation type` key work from the geometry section in each DA application, following recent changes in [PR#1002](https://github.com/JCSDA-internal/mpas-jedi/pull/1002) to use a generic interpolator for resolution changes.  The build is updated for this new change in mpas-jedi. Here, the `&GetValues` anchors and references (`*GetValues`) in the observation yaml files for the unstructured interpolation type with `nnereast: 3`( for the 3 nearest cells) is also removed. This was removed long time ago  in [PR#877](https://github.com/JCSDA-internal/mpas-jedi/pull/877).

### Issue closed
Closes #334 

### Tests completed
#### Tier 1:
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_IAU_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart
 - [x] getkf_OIE120km_WarmStart
 - [x] ForecastFromGFSAnalysesMPT