Remove job specification for forecast entry in scenarios and verifyobs from default list in forecast post (#345)
This PR removes all jobs specification for the `forecast` entry in the scenario YAML files under the `scenarios` folder. It also removes some unnecessary specifications for the `variational` entry (adjusting the default value in the case for `3dhybrid`). In addition, it removes `verifyobs` from the default list in the `post` for forecast.

Scenarios with IASI seemed out-of-date. Job specification needs to be added in case differs from default.

None

 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_IAU_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart
 - [x] getkf_OIE120km_WarmStart
 - [x] ForecastFromGFSAnalysesMPT