Remove job specification for forecast entry in scenarios and verifyobs from default list in forecast post (#350)
## Description

Similar to PR [345](https://github.com/NCAR/MPAS-Workflow/pull/345) to remove all jobs specification for the `forecast` entry in the scenario YAML files under the `scenarios` folder and `verifyobs` from the default list in the `post` for forecast. It also removes two scenarios: for abiahi and amsua, that were merged by mistake before.